# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/user": {
            "post": {
                "description": "Create a new user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Create a new user.",
                "parameters": [
                    {
                        "description": "Create User Request Body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.CreateUserRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user.CreateUserResponseBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/user.ErrorResponseBody"
                        }
                    }
                }
            }
        },
        "/user/{username}/transactions": {
            "get": {
                "description": "Get transactions of user's wallets sorted by newest.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Get transactions of user's wallets sorted by newest.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "username",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user.TransactionsResponseBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/user.ErrorResponseBody"
                        }
                    }
                }
            }
        },
        "/user/{username}/wallets": {
            "get": {
                "description": "Get balances of user's wallets.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Get balances of user's wallets.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "username",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user.GetWalletsResponseBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/user.ErrorResponseBody"
                        }
                    }
                }
            }
        },
        "/wallet": {
            "post": {
                "description": "Create a new wallet for user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wallet"
                ],
                "summary": "Create a new wallet for user.",
                "parameters": [
                    {
                        "description": "Create Wallet Request Body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.CreateWalletRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user.CreateWalletResponseBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/user.ErrorResponseBody"
                        }
                    }
                }
            }
        },
        "/wallet/{wallet_id}/deposit": {
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Deposit to wallet",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wallet"
                ],
                "summary": "Deposit to wallet",
                "parameters": [
                    {
                        "type": "string",
                        "description": "wallet id",
                        "name": "wallet_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Create Deposit Request Body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.DepositRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user.DepositResponseBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/user.ErrorResponseBody"
                        }
                    }
                }
            }
        },
        "/wallet/{wallet_id}/transfer": {
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Transfer to another wallet.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wallet"
                ],
                "summary": "Transfer to another wallet.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "wallet id",
                        "name": "wallet_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Create Transfer Request Body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.TransferRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user.TransferResponseBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/user.ErrorResponseBody"
                        }
                    }
                }
            }
        },
        "/wallet/{wallet_id}/withdraw": {
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Withdraw from wallet",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wallet"
                ],
                "summary": "Withdraw from wallet",
                "parameters": [
                    {
                        "type": "string",
                        "description": "wallet id",
                        "name": "wallet_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Create Withdraw Request Body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.WithdrawRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user.WithdrawResponseBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/user.ErrorResponseBody"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "github_com_cryptonlx_crypto_src_controllers_mux_user.User": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "username": {
                    "type": "string",
                    "example": "user1"
                }
            }
        },
...
        "user.CreateUserRequestBody": {
            "type": "object",
            "properties": {
                "username": {
                    "type": "string",
                    "example": "user1"
                }
            }
        },
        "user.CreateUserResponseBody": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/user.CreateUserResponseData"
                },
                "error": {
                    "type": "string",
                    "x-nullable": true,
                    "example": ""
                }
            }
        },
        "user.CreateUserResponseData": {
            "type": "object",
            "properties": {
                "user": {
                    "$ref": "#/definitions/user.CreatedUser"
                }
            }
        },
        "user.CreateWalletRequestBody": {
            "type": "object",
            "properties": {
                "currency": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "user.CreateWalletResponseBody": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/user.CreateWalletResponseData"
                },
                "error": {
                    "type": "string",
                    "x-nullable": true,
                    "example": ""
                }
            }
        },
        "user.CreateWalletResponseData": {
            "type": "object",
            "properties": {
                "wallet": {
                    "$ref": "#/definitions/user.CreatedWallet"
                }
            }
        },
        "user.CreatedUser": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 102
                },
                "username": {
                    "type": "string",
                    "example": "user1"
                }
            }
        },
...
        "user.DepositRequestBody": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "string",
                    "example": "10.23"
                },
                "nonce": {
                    "type": "integer",
                    "example": 1749286345000
                }
            }
        },
        "user.DepositResponseBody": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/user.DepositResponseData"
                },
                "error": {
                    "type": "string",
                    "x-nullable": true,
                    "example": ""
                }
            }
        },
        "user.DepositResponseData": {
            "type": "object",
            "properties": {
                "transaction": {
                    "$ref": "#/definitions/github_com_cryptonlx_crypto_src_controllers_mux_user.Transaction"
                }
            }
        },
        "user.ErrorResponseBody": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "integer",
                    "example": 0
                },
                "error": {
                    "type": "string",
                    "example": "internal_server_error"
                }
            }
        },
        "user.GetWalletsResponseBody": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/user.GetWalletsResponseData"
                },
                "error": {
                    "type": "string",
                    "x-nullable": true,
                    "example": ""
                }
            }
        },
        "user.GetWalletsResponseData": {
            "type": "object",
            "properties": {
                "user": {
                    "$ref": "#/definitions/github_com_cryptonlx_crypto_src_controllers_mux_user.User"
                },
                "wallets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_cryptonlx_crypto_src_controllers_mux_user.Wallet"
                    }
                }
            }
        },
        "user.TransactionsResponseBody": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/user.TransactionsResponseData"
                },
                "error": {
                    "type": "string",
                    "x-nullable": true,
                    "example": ""
                }
            }
        },
        "user.TransactionsResponseData": {
            "type": "object",
            "properties": {
                "transactions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_cryptonlx_crypto_src_controllers_mux_user.Transaction"
                    }
                }
            }
        },
        "user.TransferRequestBody": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "string",
                    "example": "10.23"
                },
                "destination_wallet_id": {
                    "type": "integer",
                    "example": 2
                },
                "nonce": {
                    "type": "integer",
                    "example": 1749286345000
                }
            }
        },
        "user.TransferResponseBody": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/user.TransferResponseData"
                },
                "error": {
                    "type": "string",
                    "x-nullable": true,
                    "example": ""
                }
            }
        },
        "user.TransferResponseData": {
            "type": "object",
            "properties": {
                "transaction": {
                    "$ref": "#/definitions/github_com_cryptonlx_crypto_src_controllers_mux_user.Transaction"
                }
            }
        },
        "user.WithdrawRequestBody": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "string",
                    "example": "10.23"
                },
                "nonce": {
                    "type": "integer",
                    "example": 1749286345000
                }
            }
        },
        "user.WithdrawResponseBody": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/user.WithdrawResponseData"
                },
                "error": {
                    "type": "string",
                    "x-nullable": true,
                    "example": ""
                }
            }
        },
        "user.WithdrawResponseData": {
            "type": "object",
            "properties": {
                "transaction": {
                    "$ref": "#/definitions/github_com_cryptonlx_crypto_src_controllers_mux_user.Transaction"
                }
            }
        }
    }
}